Pull out common flatfield loading code
